Task2/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose==3.3.0
psycopg2-binary==2.9.9